split off results to new component
